pyproject.toml
[tool.poetry]
name = "foodie-hotspot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
pyjwt = "^2.8.0"
django-cors-headers = "^4.3.0"
pillow = "^10.1.0"
drf-yasg = "^1.21.7"
gunicorn = "^21.2.0"
pymysql = "^1.1.0"
mysqlclient = "2.2.0"
xmltodict = "^0.13.0"
apscheduler = "^3.10.4"
django-apscheduler = "^0.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
